Organized classes into packages. Added some JUnit tests.
