+++ data/duke.txt
@@ -1,3 +1,3 @@
-T/1/go for a run 
-E/0/project meeting /2020-01-25
-D/0/return book /2020-01-27
+E/0/haha hehe /2121-12-12
+T/1/eat 
+D/0/return book /2020-01-28
