+++ src/main/java/command/Command.java
@@ -1,3 +1,7 @@
+package command;
+import task.*;
+import ui.*;
+import storage.*;
 import java.io.IOException;
 
 public abstract class Command {
