+++ src/main/java/command/DoneCommand.java
@@ -1,4 +1,8 @@
+package command;
+import ui.*;
+import storage.*;
 import java.io.IOException;
+import task.*;
 
 public class DoneCommand extends Command{
     int taskNum;
@@ -7,7 +11,7 @@ public class DoneCommand extends Command{
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws IOException {
-        tasks.getTask(taskNum - 1).isDone = true;
+        tasks.getTask(taskNum - 1).markDone();
         ui.taskDone(tasks.getTask(taskNum - 1));
         storage.save(tasks);
     }
