+++ src/main/java/storage/Storage.java
@@ -1,5 +1,7 @@
+package storage;
+
+import task.*;
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.time.LocalDate;
