+++ src/main/java/task/Task.java
@@ -1,23 +1,28 @@
+package task;
 import java.util.ArrayList;
-import java.util.Scanner;
-import java.time.LocalDate;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileWriter;
-import java.io.IOException;
 
-public class TaskList {
-    public ArrayList<Task> tasks;
+public class Task {
+    protected String description;
+    protected boolean isDone;
 
-    public TaskList(ArrayList<Task> tasks){
-        this.tasks = tasks;
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
     }
 
-    public TaskList(){
-        tasks = new ArrayList<>();
+    public String getStatusIcon() {
+        return (isDone ? "[Done] " : "[Not Done] "); //return tick or X symbols
     }
 
-    public String toStringDukeTasks() {
+    public String getStatusIconInBin(){
+        return (isDone ? "1" : "0"); //1 means done, 0 means not done
+    }
+
+    public void markDone(){
+        this.isDone = true;
+    }
+
+    public static String toStringDukeTasks(ArrayList<Task> tasks) {
         String taskString = "";
         for (Task task : tasks) {
             if (task instanceof Todo) {
@@ -36,16 +41,12 @@ public class TaskList {
         return taskString;
     }
 
-    public void add(Task task){
-        tasks.add(task);
+    public String toStringTaskstxt(){
+        return "";
     }
 
-    public void delete(int taskNum){
-        tasks.remove(taskNum);
+    @Override
+    public String toString() {
+        return getStatusIcon() + description;
     }
-
-    public Task getTask(int taskNum){
-        return tasks.get(taskNum);
-    }
-
 }
