--- src/test/java/DukeTest.java
@@ -1,3 +1,0 @@
-public class DukeTest {
-
-}
