--- text-ui-test/ACTUAL.TXT
@@ -1,69 +1,0 @@
-    ____________________________________________________________
-      Hello! I'm Duke
-      Whatcha wanna do?
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Got it. I've added this task:
-       [T][Not Done] borrow book
-     Now you have 1 tasks in the list.
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Got it. I've added this task:
-       [D][Not Done] return book (by: Sunday)
-     Now you have 2 tasks in the list.
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Got it. I've added this task:
-       [E][Not Done] project meeting (at: Mon 2-4pm)
-     Now you have 3 tasks in the list.
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Nice! I've marked this task as done:
-       [E][Done] project meeting (at: Mon 2-4pm)
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Nice! I've marked this task as done:
-       [D][Done] return book (by: Sunday)
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Nice! I've marked this task as done:
-       [T][Done] borrow book
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Here are the tasks in your list:
-      1.[T][Done] borrow book
-      2.[D][Done] return book (by: Sunday)
-      3.[E][Done] project meeting (at: Mon 2-4pm)
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Noted. I've removed this task:
-       [D][Done] return book (by: Sunday)
-     Now you have 2 tasks in the list.
-    ____________________________________________________________
-
-    ____________________________________________________________
-     Here are the tasks in your list:
-      1.[T][Done] borrow book
-      2.[E][Done] project meeting (at: Mon 2-4pm)
-    ____________________________________________________________
-
-    ____________________________________________________________
-      OOPS!!! The description of a todo cannot be empty.
-    ____________________________________________________________
-
-    ____________________________________________________________
-      OOPS!!! I'm sorry, but I don't know what that means :-(
-    ____________________________________________________________
-
-    ____________________________________________________________
-      See ya later alligator!
-    ____________________________________________________________
-
